workshop/serializers.py
from .models import Workshop, Tool
from django.contrib.auth.models import User
from rest_framework import serializers

class WorkshopSerializer(serializers.ModelSerializer):
    class Meta:
        model = Workshop
        fields = "__all__"

    def create(self, validated_data):
        """
        Create and return a new `Snippet` instance, given the validated data.
        """
        return Workshop.objects.create(**validated_data)

class ToolSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tool
        fields = "__all__"