requirements.txt
Django==5.0.7
djangorestframework
environs